refactor: create study package
- separate mission and lecture code
