+++ src/main/java/study/favorite/Application.java
@@ -1,4 +1,4 @@
-package subway;
+package study.favorite;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
