+++ src/main/java/study/subway/Application.java
@@ -1,4 +1,4 @@
-package favorite;
+package study.subway;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
